Errors solved
