+++ src/main/java/in/micheal/service/UserService.java
@@ -2,10 +2,12 @@ package in.micheal.service;
 
 import in.micheal.dao.UserDetailsDAO;
 import in.micheal.model.UserDetails;
-import in.micheal.validator.AdminLoginValidator;
-import in.micheal.validator.AdminRegistrationValidator;
+import in.micheal.validator.UserServiceValidator;
 
 public class UserService {
+	private UserService() {
+		// default constructor
+	}
 
 	/**
 	 * This methods return true if the log in credentials is true
@@ -16,7 +18,7 @@ public class UserService {
 	 */
 	public static boolean adminLogin(long adminId, String password) {
 		boolean verification = false;
-		boolean confirmation = AdminLoginValidator.loginValidator(adminId, password);
+		boolean confirmation = UserServiceValidator.AdminloginValidator(adminId, password);
 		if (confirmation) {
 			verification = true;
 		}
@@ -32,7 +34,7 @@ public class UserService {
 	 */
 	public static boolean adminRegistration(UserDetails adminObj) {
 		boolean confirmation = false;
-		boolean adminIdRepeatation = AdminRegistrationValidator.adminRegistrationValidator(adminObj);
+		boolean adminIdRepeatation = UserServiceValidator.adminRegistrationValidator(adminObj);
 
 		if (!adminIdRepeatation) {
 			UserDetailsDAO.addUser(adminObj);
